+++ tests/Integration/Queue/ResumableJobTest.php
@@ -30,38 +30,6 @@ class ResumableJobTest extends QueueTestCase
         //$app['config']->set('queue.default', 'database');
     }
 
-    protected function setUp(): void
-    {
-        StateHolder::$data = [];
-        parent::setUp();
-    }
-
-    public function test_workflow()
-    {
-        // @todo move this to a unit test
-        $workflow = $this->app->make(JobSequence::class);
-        $store = new ArrayStore();
-        $workflow->step(function (ExecutionStateOG $state) {
-            $state->set('hello', 'world');
-        })->step(function (ExecutionStateOG $state) {
-            $state->name = 'luke';
-        }, 'step2')->step(function (ExecutionStateOG $state) {
-            $state->name = 'taylor';
-        });
-        $workflow->persistenceCallback(fn (ExecutionStateOG $state) => $store->put('resume', $state, 100));
-        $workflow->clearStateCallback(function ($state) use ($store) {
-            $this->assertEquals([
-                'hello' => 'world',
-                'name' => 'taylor',
-            ], $state->data());
-            $this->assertEquals(3, $state->stepIndex);
-            $store->forget('resume');
-        });
-        $workflow->execute();
-
-        $this->assertEmpty($store->all());
-    }
-
     public function test_job()
     {
         /*
@@ -75,39 +43,6 @@ class ResumableJobTest extends QueueTestCase
          *
          */
     }
-
-    public function test_dispatchedJob()
-    {
-        TestResumableJob::dispatch();
-        $this->assertCount(2, StateHolder::$data);
-        [$resumeState1, $resumeState2] = StateHolder::$data;
-        $this->assertInstanceOf(ExecutionStateOG::class, $resumeState1);
-        $this->assertSame(0, $resumeState1->stepIndex);
-        $this->assertSame([], $resumeState1->data());
-        $this->assertInstanceOf(ExecutionStateOG::class, $resumeState2);
-        $this->assertSame(1, $resumeState2->stepIndex);
-        $this->assertSame(['abc' => 123, 'xyz' => 456], $resumeState2->data());
-        $this->assertEmpty(DB::table('cache')->get());
-    }
-
-    public function testRepeatingStep()
-    {
-        TestRepeatingStepResumableJob::dispatch();
-
-        $this->assertCount(3, StateHolder::$data);
-        $this->assertSame(0, StateHolder::$data[0]->stepIndex);
-        $this->assertSame([], StateHolder::$data[0]->data());
-        $this->assertSame(1, StateHolder::$data[1]->stepIndex);
-        $this->assertSame(['abc' => 123, 'xyz' => 456], StateHolder::$data[1]->data());
-        $this->assertSame(2, StateHolder::$data[2]->stepIndex);
-        $this->assertSame(['abc' => 123, 'xyz' => 456], StateHolder::$data[2]->data());
-    }
-}
-
-class StateHolder
-{
-    /** @var list<ExecutionStateOG> */
-    public static array $data;
 }
 
 class TestResumableJob implements ShouldQueue, Resumable
@@ -118,110 +53,6 @@ class TestResumableJob implements ShouldQueue, Resumable
 
     public function handle()
     {
-        $this->sequence
-            ->step(function (ExecutionStateOG $state) {
-                StateHolder::$data[$state->stepIndex] = clone $state;
-                $state->set('abc',  123);
-            })->step(function (ExecutionStateOG $state) {
-                $state->set('xyz', 456);
-                StateHolder::$data[$state->stepIndex] = clone $state;
-            }, 'step2');
-    }
-}
-
-class TestRepeatingStepResumableJob extends TestResumableJob
-{
-    public function handle(): void
-    {
-        parent::handle();
-        $this->sequence->step(name: 'step2');
-    }
-}
-
-class CheckForUpdate implements ShouldQueue, Resumable
-{
-    use InteractsWithQueue;
-    use ResumableTrait;
-    use Dispatchable;
-
-    public function handle(): void
-    {
-        $this->sequence
-            // The state of the job is persisted in cache after each step
-            // if there's a failure or an interrupt, the job will requeue
-            // and start from the failure.
-            ->step($this->getData(...))
-            ->step($this->persistData(...))
-            ->step($this->sendEmail(...));
-    }
-
-    private function getData(ExecutionStateOG $resumeState): void
-    {
-        $resumeState->data()['response'] = Http::get('https://jobs.laravel.com/jobs/')->json();
-    }
-
-    private function persistData(ExecutionStateOG $resumeState): void
-    {
-        foreach($resumeState->data['response']['data'] as $jobData) {
-            DB::table('laravel_jobs')->insertGetId([
-                'id' => $jobData['id'],
-                'data' => json_encode($jobData),
-            ]);
-        }
-    }
-
-    private function sendEmail(ExecutionStateOG $resumeState): void
-    {
-        Mail::raw("New jobs\n", fn ($message) => $message
-            ->to('luke@kuzmish.com')
-            ->subject('Demo Email')
-        );
-    }
-}
-
-class LaravelJobNotification implements Mailable
-{
-    public function __construct($value)
-    {
-        $this->value = $value;
-    }
-    public function send($mailer)
-    {
-        // TODO: Implement send() method.
-    }
-
-    public function queue(Factory $queue)
-    {
-        // TODO: Implement queue() method.
-    }
-
-    public function later($delay, Factory $queue)
-    {
-        // TODO: Implement later() method.
-    }
-
-    public function cc($address, $name = null)
-    {
-        // TODO: Implement cc() method.
-    }
-
-    public function bcc($address, $name = null)
-    {
-        // TODO: Implement bcc() method.
-    }
-
-    public function to($address, $name = null)
-    {
-        // TODO: Implement to() method.
-    }
-
-    public function locale($locale)
-    {
-        // TODO: Implement locale() method.
-    }
-
-    public function mailer($mailer)
-    {
-        // TODO: Implement mailer() method.
+        // @todo
     }
 }
